refactor (LottoMachine): refactor codes(2)
