+++ src/main/kotlin/lotto/Main.kt
@@ -7,22 +7,27 @@ import lotto.model.Money
 import lotto.model.WinningLotto
 import lotto.model.LottoNumber
 import lotto.model.Result
+import lotto.view.InputReader
 import lotto.view.InputView
 import lotto.view.ResultView
 import java.math.BigDecimal
 
 fun main() {
     val lottoMachine = LottoMachine()
+    val inputReader = InputReader()
 
-    val budget: Money = InputView.askBudgetQuestion()
+    InputView.printBudgetQuestion()
+    val budget: Money = inputReader.readBudget()
     lottoMachine.insertMoney(budget)
-    ResultView.printLottoCount(lottoMachine)
-
     val myLottos: Lottos = lottoMachine.buy()
+    ResultView.printLottoCount(lottoMachine)
     ResultView.printMyLottos(myLottos)
 
-    val winningNumbers: Lotto = InputView.askWinningLottoNumbers()
-    val bonusNumber: LottoNumber = InputView.askBonusNumber(winningNumbers)
+    InputView.printWinningNumberQuestion()
+    val winningNumbers: Lotto = inputReader.readWinningLottoNumbers()
+
+    InputView.printBonusBallQuestion()
+    val bonusNumber: LottoNumber = inputReader.readBonusNumber(winningNumbers)
     val result: List<Result> = lottoMachine.getResult(WinningLotto(winningNumbers, bonusNumber))
     val earningRate: BigDecimal = lottoMachine.getEarningRate()
     ResultView.printResult(result)
